Fixed Product table categoryID froeign key issue

+++ Core/Model/Category.cs
@@ -5,7 +5,7 @@ namespace Core.Model
     public class Category
     {
         [Key]
-        public int CategoryId { get; set; }
+        public long CategoryId { get; set; }
         [Required]
         [MaxLength(50)]
         public string CategoryName { get; set; }
